--- types/types.d.ts
@@ -1,16 +1,0 @@
-type StakingPool = {
-    Id: number;
-    apy: number;
-    lockDuration: number;
-  };
-  
-  type UserStakingInfo = {
-    claimableTokens: number;
-    lastUpdate: number;
-    stakingDuration: number;
-    stakingPool: StakingPool;
-    stakingStartDate: number;
-    totalStaked: number;
-    userId: string;
-    walletAddress: string;
-  };
